recyclerview gestures + preference manager

+++ app/src/main/java/com/ekdorn/silentium/models/Contact.kt
@@ -1,4 +1,4 @@
 package com.ekdorn.silentium.models
 
 
-data class Contact(val name: String?, val contact: String, val wasOnline: Int?)
+data class Contact(val name: String?, val contact: String, val wasOnline: Long?)
